LayoutTests/media/video-replaces-poster-expected.txt

Test for https://bugs.webkit.org/show_bug.cgi?id=34966. 
You should see the video below.

EXPECTED (1 == '1') OK
